accessible/tests/mochitest/text/test_singleline.html

<!DOCTYPE html>
<html>
<head>
  <title>nsIAccessibleText getText related function tests for html:input,html:div and html:textarea</title>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />

  <script type="application/javascript"
          src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <script type="application/javascript"
          src="../common.js"></script>
  <script type="application/javascript"
          src="../text.js"></script>
  <script type="application/javascript">
    function doTest()
    {
      // __h__e__l__l__o__ __m__y__ __f__r__i__e__n__d__
      //  0  1  2  3  4  5  6  7  8  9 10 11 12 13 14 15

      ////////////////////////////////////////////////////////////////////////
      // getTextAfterOffset

      var IDs = [ "input", "div", "editable", "textarea" ];
      var regularIDs = [ "input", "div", "editable" ];

      // BOUNDARY_LINE_START
      testTextAfterOffset(0, BOUNDARY_LINE_START, "", 15, 15, IDs);
      testTextAfterOffset(1, BOUNDARY_LINE_START, "", 15, 15, IDs);
      testTextAfterOffset(14, BOUNDARY_LINE_START, "", 15, 15, IDs);
      testTextAfterOffset(15, BOUNDARY_LINE_START, "", 15, 15, IDs);

      // BOUNDARY_LINE_END
      testTextAfterOffset(0, BOUNDARY_LINE_END, "", 15, 15, IDs);
      testTextAfterOffset(1, BOUNDARY_LINE_END, "", 15, 15, IDs);
      testTextAfterOffset(14, BOUNDARY_LINE_END, "", 15, 15, IDs);
      testTextAfterOffset(15, BOUNDARY_LINE_END, "", 15, 15, IDs);

      ////////////////////////////////////////////////////////////////////////
      // getTextBeforeOffset

      var IDs = [ "input", "div", "editable", "textarea" ];

      // BOUNDARY_LINE_START
      testTextBeforeOffset(0, BOUNDARY_LINE_START, "", 0, 0, IDs);
      testTextBeforeOffset(1, BOUNDARY_LINE_START, "", 0, 0, IDs);
      testTextBeforeOffset(14, BOUNDARY_LINE_START, "", 0, 0, IDs);
      testTextBeforeOffset(15, BOUNDARY_LINE_START, "", 0, 0, IDs);

      // BOUNDARY_LINE_END
      testTextBeforeOffset(0, BOUNDARY_LINE_END, "", 0, 0, IDs);
      testTextBeforeOffset(1, BOUNDARY_LINE_END, "", 0, 0, IDs);
      testTextBeforeOffset(14, BOUNDARY_LINE_END, "", 0, 0, IDs);
      testTextBeforeOffset(15, BOUNDARY_LINE_END, "", 0, 0, IDs);

      ////////////////////////////////////////////////////////////////////////
      // getTextAtOffset

      IDs = [ "input", "div", "editable", "textarea" ];
      regularIDs = [ "input", "div", "editable" ];

      // BOUNDARY_LINE_START
      testTextAtOffset(0, BOUNDARY_LINE_START, "hello my friend", 0, 15, IDs);
      testTextAtOffset(1, BOUNDARY_LINE_START, "hello my friend", 0, 15, IDs);
      testTextAtOffset(14, BOUNDARY_LINE_START, "hello my friend", 0, 15, IDs);
      testTextAtOffset(15, BOUNDARY_LINE_START, "hello my friend", 0, 15, IDs);

      // BOUNDARY_LINE_END
      testTextAtOffset(0, BOUNDARY_LINE_END, "hello my friend", 0, 15, IDs);
      testTextAtOffset(1, BOUNDARY_LINE_END, "hello my friend", 0, 15, IDs);
      testTextAtOffset(14, BOUNDARY_LINE_END, "hello my friend", 0, 15, IDs);
      testTextAtOffset(15, BOUNDARY_LINE_END, "hello my friend", 0, 15, IDs);

      SimpleTest.finish();
    }

    SimpleTest.waitForExplicitFinish();
    addA11yLoadEvent(doTest);
  </script>
</head>
<body>

  <a target="_blank"
     title="nsIAccessibleText getText related function tests for html:input,html:div and html:textarea"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=452769">
    Bug 452769
  </a>
  <a target="_blank"
     title="getTextAtOffset for word boundaries: beginning of a new life"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=853340">
    Bug 853340
  </a>
  <a target="_blank"
     title="getTextBeforeOffset for word boundaries: evolving"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=855732">
    Bug 855732
  </a>
  <a target="_blank"
     title=" getTextAfterOffset for line boundary on new rails"
     href="https://bugzilla.mozilla.org/show_bug.cgi?id=882292">
    Bug 882292
  </a>
  <p id="display"></p>
  <div id="content" style="display: none"></div>
  <pre id="test">
  </pre>

  <input id="input" value="hello my friend"/>
  <div id="div">hello my friend</div>
  <div id="editable" contenteditable="true">hello my friend</div>
  <textarea id="textarea">hello my friend</textarea>

</body>
</html>
